feat(poc): memory optimization - malloc_trim + VisitedList uint8 + streaming adjacency0 free
Memory improvements (DEC-063 promote prerequisite):
- malloc_trim(0) after CSR build + UpperPQ free: returns memory to OS
- VisitedList uint32 -> uint8: 10M nodes saves 30MB
- Streaming adjacency0 free: release inner vectors per-node during CSR build,
  reduces peak from ~3.5GB to ~2.7GB

Results (no cgroup):
- RSS init: 2422MB -> 1792MB (-630MB)
- RSS peak: 2483MB -> 1923MB (-560MB)

3GB cgroup now works (was OOM at 4GB before):
- R0: 56.0 QPS, R1: 151.1 QPS (+169.8%)
- Tighter memory -> more I/O pressure -> bigger pipe_ring_ benefit
- Recall: 94.85% (unchanged)

cgroup peak: 3072MB (3GB limit, memory pressure during graph loading)

+++ poc/io-pipelining/disk_hnsw_pipe.h
@@ -51,8 +51,8 @@ struct PQParams {
 // VisitedList: 简单的访问标记数组（不需要池化，每次搜索创建一个）
 // ============================================================
 struct VisitedList {
-    std::vector<uint32_t> mass;  // 使用uint32_t作为标记类型
-    uint32_t curV;               // 当前标记值
+    std::vector<uint8_t> mass;   // uint8: 10M 节点省 30MB vs uint32 (DEC-063 P2 优化)
+    uint8_t curV;                // 当前标记值
 
     explicit VisitedList(size_t num_elements)
         : mass(num_elements, 0), curV(1) {}
